math/pyproject.toml
[project]
name = "math-agent"
version = "0.1.0"
description = "A Math Agent using Gemini"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk[http-server]",
    "google-generativeai",
    "python-dotenv",
    "uvicorn",
    "requests",
]
